Integrates React Query for data fetching
Sets up React Query for managing and caching API data.
Introduces a `useGetUser` hook for fetching user data, demonstrating the
integration. The API calls in the client component are updated to utilize
the new data fetching hook.

Removes the direct API call in the client component
and replaces it with the `useGetUser` hook.

Creates an API proxy in next.config.js to prevent CORS issues.